Huh?
git-svn-id: file:///p/wpis/people/students/kidd/svnwali/trunk@1400 dfa36f5b-b2fa-0310-9605-e5aca7e64462

+++ Source/wali/graph/RegExp.cpp
@@ -385,7 +385,7 @@ namespace wali {
                     next_it++;
                     if(next_it == heap.end())
                         break;
-                    if( (int)((*it)->outnodes.size() + (*next_it)->outnodes.size()) < min) {
+                    if( (*it)->outnodes.size() + (*next_it)->outnodes.size() < min) {
                         min_pos = it;
                         min = (*it)->outnodes.size() + (*next_it)->outnodes.size();
                     }
